Rename Examples/Xunit to Examples/TestingPatterns; fix Showcase README staleness
The folder name 'Xunit' described the *runner* not the *intent*. Renamed
to 'TestingPatterns' so the sample's purpose (FluentValidation, Maybe,
async, parallel, DDD test patterns) is self-evident from the folder name.

- Examples/Xunit/                        -> Examples/TestingPatterns/
- Example.Tests.csproj                   -> TestingPatterns.Tests.csproj
  (.Tests suffix preserved so Directory.Build.props IsTestProject detection
   still kicks in)
- 'namespace Example.Tests'              -> 'namespace TestingPatterns'
  (in source files; also fixed fully-qualified Example.Tests.* references
   in FluentValidationSamplesTests.cs)
- Trellis.slnx project path updated.

Showcase README: drop reference to deleted AuthorizationExample; point at
InMemoryIdentityVerifier instead for the Forbidden-with-policyId pattern.

All 24 Showcase tests still pass; TestingPatterns tests still pass.
