+++ Examples/TestingPatterns/AsyncUsageExamples.cs
@@ -1,4 +1,4 @@
-﻿namespace Example.Tests;
+namespace TestingPatterns;
 
 using System.Diagnostics;
 using Trellis;
