+++ Examples/TestingPatterns/MULTI_STAGE_PARALLEL_EXAMPLES.md
