+++ Examples/TestingPatterns/TestingPatterns.Tests.csproj
