+++ Examples/TestingPatterns/TraceFixture.cs
@@ -1,4 +1,4 @@
-﻿namespace Example.Tests;
+namespace TestingPatterns;
 
 using System;
 using System.Diagnostics;
