+++ Trellis.slnx
@@ -39,7 +39,7 @@
     <Project Path="Examples/EfCoreExample/EfCoreExample.csproj" />
     <Project Path="Examples/ConditionalRequestExample/ConditionalRequestExample.csproj" />
     <Project Path="Examples/SsoExample/SsoExample.csproj" />
-    <Project Path="Examples/Xunit/Example.Tests.csproj" />
+    <Project Path="Examples/TestingPatterns/TestingPatterns.Tests.csproj" />
   </Folder>
   <Folder Name="/Examples/Showcase/">
     <Project Path="Examples/Showcase/src/Showcase.Domain/Showcase.Domain.csproj" />
